+++ src/Zef/Mini/IMarshaller.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Zef\Mini;
+
+interface IMarshaller
+{
+    public function marshal( $row);
+    
+    public function marshalAll( $data);
+    
+}